new: role crud

+++ src/main/java/ru/bash/jopka/Controller/RoleController.java
@@ -44,7 +44,7 @@ public class RoleController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@Positive @PathVariable int id) {
+    public ResponseEntity<String> delete(@Positive @PathVariable long id) {
         return new ResponseEntity<>(service.delete(id), HttpStatus.OK);
     }
 }
